TESTED - KA: MotionProfileOnboardRunner attempted
MotionProfileOnboardRunner created to allow us to do motion profiling on
the Rio instead of on the Talons so we can utilize the gyro. It's worked
on both Janderson and Kongo, but doesn't work on Delta. Old motion
profile autos should still work, but they haven't been tested since the
changes.